requirements.txt
streamlit
streamlit-webrtc
paddleocr
paddlepaddle
pillow
numpy

